[profiler] Rename proflog.{c,h} to mono-profiler-log.{c,h} for consistency with other profilers.
